+++ CHANGELOG.md
@@ -6,6 +6,7 @@ More infos about this file : http://keepachangelog.com/
 
 ## [Unreleased] - no_due_date
 
+- **Fix english naming: `Ambigue` -> `Ambiguous`**
 - **DistanceApi#distances method should also accept `Array`s**
 - **Some small optimisations**